Update Denial Level
The tiles have been set up and put into the scene. I need to add some variation and grass to the foreground of the tile map. The background and far background is in and parallax needs to be tested. Created a game object for falling platforms that needs to be finished.