Denial Level Update
All backgrounds have been inserted. The parallax layer was adjusted and the boss room expanded. The boss room will need to be further tested to ensure the background looks okay. I have memento placeholders in the level and attached to update with the UI. Press F to pick up mementos. A basic phantom path ability was created and attached to both the player controller and the UI controller. It could use some tweaking to ensure it works smoothly. I am still working on getting the falling platforms and puzzles into the level. I will also create a door to close behind the player when they enter the boss room.

We still need to create the bark and sniff ability and ensure it properly interacts with enemies and the environment. We also need to add enemies into this level.