Update Denial Level
I added a trampoline that can be bounced on, although the player controller seems to interfere with it, so I made some changes. It will need to be tweaked. I added in some boxes for the puzzle mechanic. Overall, the obstacles/trampolines/phantom path ability are a little funky, but they work okay for testing purposes. The collision box on the player needs to be fixed. I will finish editing the level to match the new player mechanics tomorrow.