Added Sniff Ability & Outlined Anger Level
I added a basic sniff ability. Currently, it just changes the color of falling platforms to red for a couple seconds and mementos to yellow for a couple seconds. This could probably be improved later by having the artists draw outlines for the object, but it works for now. I also started outlining the anger level.