Major Updates
I changed the music on the menu screen. I added the anger bark ability and it created a breakable wall object that it works with. I still need to create the falling objects. I started working on the trigger/door mechanic for when the player enters the boss room. I added a basic enemy health script, but this should only be applied starting in the anger level to interact with the anger ability.