stalactites & sliding
The player character now slides on sloped areas, but is still easily able to spam jump to go back up. That will need to be fixed later. Stalactites react to the shatter bark ability, their sprites change, and they fall to the ground. If it hits solid ground the player can step on it, but it will not move. If it falls into lava, it hits the top and slowly sinks. I need to tweak the sinking as sometimes it hits the lava and disappears and only sinks if the player is on it. Also, I might need to convert the lava areas to game object that ignore the player. Currently, I'm using the foreground and had to add a tilemap for the stalactite to hit, but the player can now also stand on lava.