Boss Fight
I added a trigger and stopping point for some lava in the boss room. I also created a water splash object that will finish the boss fight, but I still need to connect it to the boss/lava to solidify those when it and add a stalactite that appears when all armor is off to bark at before spraying water down.